getBundleResource now locks.
Fixed typo in log message for loadBundle.

loadBundle now locks on the write lock and attempts to uninstall old bundles when installing a new one.

getPlugin locks on read.

When checking to load bundles, now compares the creation time of the bundle to possibly load and the timestamp of the existing bundle as Instants rather than their milliseconds as longs.