Disabled all caching for plugin resources.
This fixes a bug where if a request gets a plugin resource, said plugin is removed and it attempts to re-request the resource with caching enabled, it will ask Spring to validate the cache, Spring will try to access the URL of the bundle which no longer exists and fail with an exception.

Possibly an opportunity for a custom caching validator implementation.