Fixed the plugin manager crapping out when it tried to load bundles with the same id from multiple sources as a time.
Bundles are uniquely identified by a string path. Previously, the absolute path of the URL was used, so mixing loading from both the file system and S3 would mean that the location strings wouldn't match so the framework would try to load the "new" plugin but blowup because it would have the same name and id combination OSGi uses to determine uniqueness.